// src/types/partnerPreferences.ts
export type PartnerPreferences = {
  id: 1;
  gender: string[];
  minAge: number;
  maxAge: number;
  bodyType: string;
  commitmentLevel: string;
  personalBelieves: string[];
  favoriteHolidayDestination: string;
  freeTime: string[];
  musicGenres: string[];
  pet: string;
  userId: number;
};